requirements.txt
streamlit==1.25.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy<2,>=1.19.3
pandas==2.2.2
protobuf<5,>=4.25.3
pyvista
pyvistaqt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client